added custom python c module

+++ dict/hashutils.c
@@ -1,7 +1,7 @@
 #include "headers/hash.h"
 #include <string.h>
 
-uint64_t index_from_hash(const int hash, const int size)
+uint64_t index_from_hash(const uint64_t hash, const size_t size)
 {
     return hash & (size - 1);
 }
